Update to nikic/php-parser 4.13.0 (#904)
* Update to nikic/php-parser 4.13.0

* check Arg on PregMatchTypeCorrector

* [ci-review] Rector Rectify

* [ci-review] Rector Rectify

* check Aeg on NodeTransformer

* check arg on RegexPatternArgumentManipulator

* [ci-review] Rector Rectify

* check Arg on FuncCallManipulator

* check Arg on ClassMethodAssignManipulator

* check Arg on StringifyStrNeedlesRector

* [ci-review] Rector Rectify

* Arg check on CompactFuncCallAnalyzer

* Arg check on CompactFuncCallAnalyzer

* check Arg on CallableInMethodCallToVariableRector

* check Arg on CallableInMethodCallToVariableRector

* Arg check on BeConstructedWithAssignFactory

* Add ArgsAnalyzer to check Arg instance in in Args position

* Arg check on SubstrMatchAndRefactor

* Arg check on NewArgToMethodCallRector

* check Arg on ConsistentImplodeRector

* Arg check on DuringMethodCallFactory

* Arg check on CountOnNullRector

* Arg check on ExportToReflectionFunctionRector

* Arg check on ImplodeAnalyzer

* Arg check on WhileEachToForeachRector

* Arg check on DowngradeIsCountableRector

* Arg check on VariableNaming

* Arg check on ConditionResolver

* Arg check on DirNameFileConstantToDirConstantRector

* Arg check on NewFluentChainMethodCallNodeAnalyzer

* Arg check on MbStrrposEncodingArgumentPositionRector

* Arg check on UnwrapFutureCompatibleIfFunctionExistsRector

* Arg check on JsonArrayFactory

* Arg check on RemoveFuncCallRector

* Arg check on ArrayKeyExistsOnPropertyRector

* Arg check on FunctionExistsFunCallAnalyzer

* Arg check on SetTypeToCastRector

* Arg check on ReplaceEachAssignmentWithKeyCurrentRector

* Arg check on PregMatchTypeCorrector

* Arg check on SwapFuncCallArgumentsRector

* Arg check on FilterVarToAddSlashesRector

* Arg check on ClassOnObjectRector

* Arg check on ClassOnObjectRector

* Arg check on CallerParamMatcher

* Arg check on ArgumentFuncCallToMethodCallRector

* add ArgsAnalyzer::isArgsInstanceInArgsPositions() for multi check directly

* Arg check on ArraySpreadInsteadOfArrayMergeRector

* Arg check on GetDebugTypeRector

* Arg check on GetDebugTypeRector

* Arg check on StringifyDefineRector

* Arg check on DowngradeArrayKeyFirstLastRector

* Arg check on IsArrayAndDualCheckToAble

* Arg check on CreateFunctionToAnonymousFunctionRector

* Arg check on AddArrayDefaultToArrayPropertyRector

* Arg check on IsObjectOnIncompleteClassRector

* Arg check on ForToForeachRector

* Arg check on ClassMethodNodeRemover

* Arg check on CompactConverter

* Arg check on IsAWithStringWithThirdArgumentRector

* Arg check on CallUserFuncArrayToVariadicRector

* Arg check on MysqlPConnectToMysqliConnectRector

* Arg check on MysqlAssignToMysqliRector

* Arg check on InArrayAndArrayKeysToArrayKeyExistsRector

* Arg check on NonVariableToVariableOnFunctionCallRector

* Arg check on VersionCompareFuncCallToConstantRector

* fix test

* Arg check on StrEndsWithRector

* Arg check on TokenManipulator

* Arg check on StrStartsWithFactory

* Arg check on DowngradeIsIterableRector

* Arg check on DowngradeStripTagsCallWithArrayRector

* [ci-review] Rector Rectify

* [ci-review] Rector Rectify

* Arg check on ChangeDiffForHumansArgsRector

* Arg check on ThrowWithPreviousExceptionRector

* [ci-review] Rector Rectify

* [ci-review] Rector Rectify

* Arg check on StrposMatchAndRefactor

* Arg check on UseMessageVariableForSprintfInSymfonyStyleRector

* [ci-review] Rector Rectify

* [ci-review] Rector Rectify

* fix

* Arg check on PhpSpecClassToPHPUnitClassRector

* Arg check on CallTypesResolver

* Arg check on NonVariableToVariableOnFunctionCallRector

* Arg check on DispatchStringToObjectRector

* Arg check on GetClassToInstanceOfRector

* Arg check on StrncmpMatchAndRefactor

* Arg check on StrncmpMatchAndRefactor

* Arg check on ArraySpreadInsteadOfArrayMergeRector

* Arg check on ArrayKeysAndInArrayToArrayKeyExistsRector

* Arg check on DowngradeSessionStartArrayOptionsRector

* Arg check on CountArrayToEmptyArrayComparisonRector

* Arg check on DowngradeClosureFromCallableRector

* Arg check on EregToPregMatchRector

* Arg check on GetClassOnNullRector

* Arg check on use RemoveSoleValueSprintfRector

* Arg check on UnSpreadOperatorRector

* Arg check on DowngradeStrContainsRector

* [ci-review] Rector Rectify

* Arg check on ArrayMergeOfNonArraysToSimpleArrayRector

* Arg check on IntvalToTypeCastRector

* Arg check on RemoveUnusedVariableAssignRector

* Arg check on ConsistentPregDelimiterRector

* Arg check on CountManipulator

* Arg check on DowngradeDefineArrayConstantRector

* Arg check on RemoveReferenceFromCallRector

* Arg check on SimplifyInArrayValuesRector

* Arg check on SingleInArrayToCompareRector

* Arg check on SimplifyIfIssetToNullCoalescingRector

* Arg check on StrlenZeroToIdenticalEmptyStringRector

* [ci-review] Rector Rectify

* Add ComplexType check

* Arg check on UnwrapSprintfOneArgumentRector

* Arg check on CurrentAndParentClassMethodComparator

* Arg check on CompactToVariablesRector

* [ci-review] Rector Rectify

* ComplexType check

* Arg check on MockVariableCollector

* re-generate preload.php

* [ci-review] Rector Rectify

* fix phpstan ignore error

* Arg check on FormerNullableArgumentToScalarTypedRector

* ComplexType check on MakeTypedPropertyNullableIfCheckedRector

* Arg check on ArgumentRemoverRector

* using getArgs() on PhpSpecPromisesToPHPUnitAssertRector

* using getArgs() on args using in method call args

* Arg check on PhpSpecMocksToPHPUnitMocksRector

* Arg check on SetCookieRector

* Arg check on StringsAssertNakedRector

* Arg check on PregReplaceEModifierRector

* using getArgs() on multiple count() usage

* using getArgs() on args using in method call args

* using getArgs() on args using in method call args

* using getArgs() on args using in method call args

* Arg check on MysqlFuncCallToMysqliRector

* using getArgs() on args using in method call args

* using getArgs() on args using in method call args

* Arg check on DowngradeCovariantReturnTypeRector

* Arg check on SetCookieOptionsArrayToArgumentsRector

* using getArgs() on too complex check

* Arg check on DowngradeNegativeStringOffsetToStrlenRector

* ensure preload always used when file exists

* Arg check on SimplifyFuncGetArgsCountRector

* set back reportUnmatchedIgnoredErrors as different between local and github

* Arg check on ArgumentDefaultValueReplacer

* fix test of getArgs() on MysqlQueryMysqlErrorWithLinkRector

* Arg check on SymmetricArrayDestructuringToListRector

* Arg check on UseMessageVariableForSprintfInSymfonyStyleRector

* Arg check on ConsistentPregDelimiterRector

* Arg check on CallUserFuncToMethodCallRector

* Arg check on ArrayKeyExistsTernaryThenValueToCoalescingRector

* Arg check on SimplifyStrposLowerRector

* Arg check on CompactToVariablesRector

* Arg check on CompactToVariablesRector

* Arg check on ChangeArrayPushToArrayAssignRector

* Arg check on ChangeArrayPushToArrayAssignRector

* Arg check on CompactConverter

* final touch: Arg check on CallUserFuncWithArrowFunctionToInlineRector

* fix

+++ rules/MysqlToMysqli/Rector/FuncCall/MysqlFuncCallToMysqliRector.php
@@ -56,7 +56,7 @@ CODE_SAMPLE
     /**
      * @param FuncCall $node
      */
-    public function refactor(Node $node): FuncCall
+    public function refactor(Node $node): ?FuncCall
     {
         if ($this->isName($node, 'mysql_create_db')) {
             return $this->processMysqlCreateDb($node);
@@ -71,14 +71,14 @@ CODE_SAMPLE
             $node->args[0] = new Arg(new String_('SHOW DATABASES'));
         }
 
-        if ($this->isName($node, 'mysql_list_fields')) {
+        if ($this->isName($node, 'mysql_list_fields') && $node->args[0] instanceof Arg && $node->args[1] instanceof Arg) {
             $node->name = new Name(self::MYSQLI_QUERY);
             $node->args[0]->value = $this->joinStringWithNode('SHOW COLUMNS FROM', $node->args[1]->value);
 
             unset($node->args[1]);
         }
 
-        if ($this->isName($node, 'mysql_list_tables')) {
+        if ($this->isName($node, 'mysql_list_tables') && $node->args[0] instanceof Arg) {
             $node->name = new Name(self::MYSQLI_QUERY);
             $node->args[0]->value = $this->joinStringWithNode('SHOW TABLES FROM', $node->args[0]->value);
         }
@@ -86,16 +86,32 @@ CODE_SAMPLE
         return $node;
     }
 
-    private function processMysqlCreateDb(FuncCall $funcCall): FuncCall
+    private function processMysqlCreateDb(FuncCall $funcCall): ?FuncCall
     {
+        if (! isset($funcCall->args[0])) {
+            return null;
+        }
+
+        if (! $funcCall->args[0] instanceof Arg) {
+            return null;
+        }
+
         $funcCall->name = new Name(self::MYSQLI_QUERY);
         $funcCall->args[0]->value = $this->joinStringWithNode('CREATE DATABASE', $funcCall->args[0]->value);
 
         return $funcCall;
     }
 
-    private function processMysqlDropDb(FuncCall $funcCall): FuncCall
+    private function processMysqlDropDb(FuncCall $funcCall): ?FuncCall
     {
+        if (! isset($funcCall->args[0])) {
+            return null;
+        }
+
+        if (! $funcCall->args[0] instanceof Arg) {
+            return null;
+        }
+
         $funcCall->name = new Name(self::MYSQLI_QUERY);
         $funcCall->args[0]->value = $this->joinStringWithNode('DROP DATABASE', $funcCall->args[0]->value);
